requirements.txt
Flask==2.2.2
Flask_Login==0.6.2
flask_sqlalchemy==3.0.2
python_bcrypt==0.3.2
SQLAlchemy==1.4.46
Werkzeug==2.2.2
PyMySQL
gunicorn
